Rename backend test classes, move Settings class
Settings class is now specific to the KBase server implementation, not
the Workspace library

Strictly speaking, ShockBlobStore should also be moved to
workspace.kbase, but then the tests in workspace.test should also be
moved, and that's just too much trouble - It'd duplicate a bunch of
tests. Maybe there's some way to do it that's not nasty but it's not
worth the effort.
